Minor changes to pg_audio
Uses signed 16-bit ints for actual audio data now.

+++ src/procgl/audio.h
@@ -1,7 +1,7 @@
 #define PG_AUDIO_SAMPLE_RATE 44100
 
 struct pg_audio_chunk {
-    float* samples;
+    int16_t* samples;
     int len;
 };
